requirements.txt
py-applescript
whiptail-dialogs
configparser
